app/requirements.txt
Flask==3.1.0
python-pptx==1.0.2
openai==1.68.2